refactor: extract shared Plex sync helpers to deduplicate code (tb-189)
Extract duplicated watch-history logging and external ID extraction
logic from service.ts, sync-movies.ts, and sync-tv.ts into a shared
sync-helpers.ts module.

Shared helpers:
- extractExternalIdAsNumber: replaces 3 inline ID extraction patterns
- logMovieWatch: replaces 2 identical movie watch logging blocks
- syncEpisodeWatches: replaces 2 nearly-identical episode sync functions
